[New] Lang-Kobayashi dynamical system [Feature] Use abstract Runge-Kutta-4 function with function pointer and parameters for different dynamical systems.
The Runge-Kutta-4 (RK4) method is now written in a more generic way. This way it will be easier to implement new dynamical systems and easily integrate them using RK4.
The programm's default is now to integrate a Lang-Kobayashi-System, simplify the trajectory and safe it to a .txt file. A simple python plot.py can be used to view the result.