Major change in project.
This commit changes many parts in many files. It is the first version
that actually writes out a file holding the timeseries of various delay coupled
oscillators.

A couple of traits are defined so that dynamical systems can be written in a way
so that they can be used in a runge-kutta-4 integration process. By using traits
new dynamical systems can be added easily later on and can be plugged in.