Separate project into workspaces.
Create new workspace that contains `timeseries`-library that handles the
saving of the dynamic system's trajectories and saving it into files.
In the future timeseries analysis e.g. detection of extrema should also
be handled in this package.

The rest of the project remains in a single workspace for now. But that
will likely change.